+++ Changelog.md
@@ -16,5 +16,6 @@
 ## v0.2.0
 
 - Make the `Instrument` trait more modular and provide default implementations
+- Bake `offbeat_intensity` into the general intensity
 
 [IN PROGRESS]