+++ Changelog.md
@@ -17,5 +17,6 @@
 
 - Make the `Instrument` trait more modular and provide default implementations
 - Bake `offbeat_intensity` into the general intensity
+- Make `get_active_note()` part of `MusicTrack` trait
 
 [IN PROGRESS]